Refactor Application to MultiUpDirect and add central panel
Renamed the main Application struct to MultiUpDirect and added a central panel to the user interface. The toolbars for different options have been refactored for better clarity and easier modification. This sets up some basic functionality and gets the application to a point where it can be built and displayed.