UI fixes and API fix
Changed spacing in error log to 2 spaces. Changed width of error log to unlimited. Limited direct links output so the footer can be placed without it being interfered with. Fixed API check, so it only uses the API instead of scraping the page initially and also added direct links and status properly.